message render: Fold messageReaction*AsHtml into main file.
